backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
pandas==2.2.0
python-dotenv==1.0.1
google-genai==1.38.0
markdown-pdf==0.1.0